backend/requirements.txt
asgiref==3.7.1
astroid==2.15.5
black==23.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
dill==0.3.6
Django==4.2.1
django-cors-headers==4.0.0
django-environ==0.10.0
djangorestframework==3.14.0
gunicorn==20.1.0
idna==3.4
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.1
psycopg2-binary==2.9.6
pylint==2.17.4
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
tomli==2.0.1
tomlkit==0.11.8
typing_extensions==4.6.1
urllib3==2.0.2
whitenoise==6.4.0
wrapt==1.15.0
